Add command `aworld run`
