+++ aworld/__main__.py
@@ -1,2 +1,7 @@
 # coding: utf-8
 # Copyright (c) 2025 inclusionAI.
+
+from .web.cli import main
+
+if __name__ == "__main__":
+    main()
