+++ aworld/requirements.txt
@@ -7,3 +7,5 @@ mcp[cli]~=1.6.0
 python-dotenv~=1.0.1
 executing~=2.2.0
 kubernetes~=32.0.1
+streamlit>=1.32.0
+pyvis>=0.3.2
