--- examples/__init__.py
@@ -1,4 +1,0 @@
-# coding: utf-8
-# Copyright (c) 2025 inclusionAI.
-
-import examples.tools
