+++ examples/web/README.md
@@ -1,0 +1,19 @@
+
+
+## Run Agent via `aworld run`
+
+
+### Directory Structure
+
+The web application follows this directory structure:
+
+```
+.env
+agent_deploy
+|--agent_x
+    |--__init__.py
+    |--agent.py
+|--agent_y
+    |--__init__.py
+    |--agent.py
+```
