--- examples/web/requirements.txt
@@ -1,5 +1,0 @@
-streamlit>=1.32.0
-openai>=1.12.0
-python-dotenv>=1.0.0
-watchdog
-pyvis>=0.3.2
